Add Subscription

+++ app/src/main/java/com/kfouri/brutest/network/ApiHelper.kt
@@ -3,5 +3,5 @@ package com.kfouri.brutest.network
 class ApiHelper(private val apiService: ApiService) {
     suspend fun getMovies() = apiService.getMovies()
     suspend fun getGenres() = apiService.getGenres()
-    suspend fun getMovieDetail(id: Int) = apiService.getMovieDetail(id)
+    suspend fun getMovieDetail(id: Long) = apiService.getMovieDetail(id)
 }
