+++ app/src/main/java/com/kfouri/brutest/network/ApiRepository.kt
@@ -3,5 +3,5 @@ package com.kfouri.brutest.network
 class ApiRepository(private val apiHelper: ApiHelper) {
     suspend fun getMovies() = apiHelper.getMovies()
     suspend fun getGenres() = apiHelper.getGenres()
-    suspend fun getMovieDetail(id: Int) = apiHelper.getMovieDetail(id)
+    suspend fun getMovieDetail(id: Long) = apiHelper.getMovieDetail(id)
 }
